fix: combine all tests in one file (only for discover problems)
